refactor: decouple tool execution from mcp server
