+++ packages/server/api/src/app/mcp/mcp-server/mcp-server.ts
@@ -1,45 +1,37 @@
-import { ActionBase } from '@activepieces/pieces-framework'
-import { rejectedPromiseHandler, UserInteractionJobType } from '@activepieces/server-shared'
+import { rejectedPromiseHandler } from '@activepieces/server-shared'
 import {
     AI_USAGE_FEATURE_HEADER,
     AI_USAGE_MCP_ID_HEADER,
     AIUsageFeature,
     EngineResponseStatus,
-    ExecuteActionResponse,
     isNil,
     McpFlowTool,
     McpPieceTool,
-    McpPieceToolData,
     McpRunStatus,
     mcpToolNaming,
     McpToolType,
     McpTrigger,
-    PiecePackage,
-    spreadIfDefined,
     TelemetryEventName,
 } from '@activepieces/shared'
 import { createOpenAI } from '@ai-sdk/openai'
 import { LanguageModelV2 } from '@ai-sdk/provider'
 import { McpServer } from '@modelcontextprotocol/sdk/server/mcp.js'
-import { generateObject } from 'ai'
 import { FastifyBaseLogger } from 'fastify'
 import { StatusCodes } from 'http-status-codes'
-import { EngineHelperResponse } from 'server-worker'
-import { z, ZodRawShape } from 'zod'
+import { z } from 'zod'
 import { accessTokenManager } from '../../authentication/lib/access-token-manager'
 import { domainHelper } from '../../ee/custom-domains/domain-helper'
 import { flowService } from '../../flows/flow/flow.service'
 import { telemetry } from '../../helper/telemetry.utils'
-import { getPiecePackageWithoutArchive, pieceMetadataService } from '../../pieces/piece-metadata-service'
+import { pieceMetadataService } from '../../pieces/piece-metadata-service'
 import { projectService } from '../../project/project-service'
 import { triggerSourceService } from '../../trigger/trigger-source/trigger-source-service'
 import { WebhookFlowVersionToRun } from '../../webhooks/webhook-handler'
 import { webhookService } from '../../webhooks/webhook.service'
-import { userInteractionWatcher } from '../../workers/user-interaction-watcher'
 import { mcpRunService } from '../mcp-run/mcp-run.service'
 import { mcpService } from '../mcp-service'
 import { mcpUtils } from '../mcp-utils'
-
+import { ToolExecution } from '../tool/tool-execution'
 
 
 export async function createMcpServer({
@@ -66,6 +58,36 @@ export async function createMcpServer({
     })
     await Promise.all(addedToolPromise)
     return { server }
+}
+
+async function initializeOpenAIModel({
+    platformId,
+    projectId,
+    mcpId,
+}: {
+    platformId: string
+    projectId: string
+    mcpId: string
+}): Promise<LanguageModelV2> {
+    const model = 'gpt-4.1'
+    const baseUrl = await domainHelper.getPublicApiUrl({
+        path: '/v1/ai-providers/proxy/openai/v1/',
+        platformId,
+    })
+
+    const apiKey = await accessTokenManager.generateEngineToken({
+        platformId,
+        projectId,
+    })
+
+    return createOpenAI({
+        baseURL: baseUrl,
+        apiKey,
+        headers: {
+            [AI_USAGE_FEATURE_HEADER]: AIUsageFeature.MCP,
+            [AI_USAGE_MCP_ID_HEADER]: mcpId,
+        },
+    }).chat(model)
 }
 
 async function addPieceToServer(
@@ -103,37 +125,22 @@ async function addPieceToServer(
         toolDescription,
         toolSchema,
         async (params) => {
-            const piecePackage = await getPiecePackageWithoutArchive(
-                logger, 
-                projectId, 
-                platformId, 
-                {
-                    pieceName: pieceMetadata.name,
-                    pieceVersion: pieceMetadata.version,
-                },
-            )
             try {
-                const parsedInputs = await extractActionParametersFromUserInstructions({
-                    actionMetadata,
-                    toolPieceMetadata,
-                    userInstructions: params.instructions,
-                    piecePackage,
+                const aiModel = await initializeOpenAIModel({
                     platformId,
                     projectId,
-                    logger,
                     mcpId: mcpTool.mcpId,
                 })
-
-                const result = await userInteractionWatcher(logger)
-                    .submitAndWaitForResponse<EngineHelperResponse<ExecuteActionResponse>>({
-                    jobType: UserInteractionJobType.EXECUTE_TOOL,
+                
+                const toolExecution = new ToolExecution({ projectId, platformId, aiModel })
+                const auth = !isNil(toolPieceMetadata.connectionExternalId) ? `{{connections['${toolPieceMetadata.connectionExternalId}']}}` : undefined
+                
+                const { result, parsedInputs } = await toolExecution.execute({
+                    auth,
+                    userInstructions: params.instructions,
                     actionName: toolPieceMetadata.actionName,
                     pieceName: toolPieceMetadata.pieceName,
                     pieceVersion: toolPieceMetadata.pieceVersion,
-                    packageType: pieceMetadata.packageType,
-                    pieceType: pieceMetadata.pieceType,
-                    input: parsedInputs,
-                    projectId,
                 })
 
                 trackToolCall({ mcpId: mcpTool.mcpId, toolName: toolActionName, projectId, logger })
@@ -312,117 +319,6 @@ async function addFlowToServer(
 }
 
 
-async function initializeOpenAIModel({
-    platformId,
-    projectId,
-    mcpId,
-}: InitializeOpenAIModelParams): Promise<LanguageModelV2> {
-    const model = 'gpt-4.1'
-    const baseUrl = await domainHelper.getPublicApiUrl({
-        path: '/v1/ai-providers/proxy/openai/v1/',
-        platformId,
-    })
-
-    const apiKey = await accessTokenManager.generateEngineToken({
-        platformId,
-        projectId,
-    })
-
-    return createOpenAI({
-        baseURL: baseUrl,
-        apiKey,
-        headers: {
-            [AI_USAGE_FEATURE_HEADER]: AIUsageFeature.MCP,
-            [AI_USAGE_MCP_ID_HEADER]: mcpId,
-        },
-    }).chat(model)
-}
-
-
-
-async function extractActionParametersFromUserInstructions({
-    actionMetadata,
-    toolPieceMetadata,
-    userInstructions,    
-    piecePackage,
-    platformId,
-    projectId,
-    logger,
-    mcpId,
-}: ExtractActionParametersParams): Promise<Record<string, unknown>> {
-    const connectionReference = !isNil(toolPieceMetadata.connectionExternalId) ? `{{connections['${toolPieceMetadata.connectionExternalId}']}}` : undefined
-
-    const aiModel = await initializeOpenAIModel({
-        platformId,
-        projectId,
-        mcpId,
-    })
-
-    const actionProperties = actionMetadata.props
-    const depthToPropertyMap = mcpUtils.sortPropertiesByDependencies(actionProperties)
-
-    const extractedParameters = await Object.entries(depthToPropertyMap).reduce(
-        async (accumulatedParametersPromise, [_, propertyNames]) => {
-            const accumulatedParameters = {
-                ...(await accumulatedParametersPromise),
-                ...spreadIfDefined('auth', connectionReference),
-            }
-
-            const parameterExtractionPrompt = mcpUtils.buildParameterExtractionPrompt({
-                propertyNames,
-                userInstructions,
-            })
-
-            const propertySchemas = (await Promise.all(propertyNames.map(async propertyName => {
-                const result = await mcpUtils.buildZodSchemaForPieceProperty({
-                    property: actionProperties[propertyName],
-                    logger,
-                    input: accumulatedParameters,
-                    projectId,
-                    propertyName,
-                    actionMetadata,
-                    piecePackage,
-                })
-                return { propertyName, ...result }
-            }))).filter(({ schema }) => schema !== null)
-
-            const schemaObject: ZodRawShape = Object.fromEntries(
-                propertySchemas
-                    .map(({ propertyName, schema }) => [propertyName, schema!]),
-            )
-
-            const propertySchemaValues = propertySchemas.map(({ value }) => value).filter(value => value !== null)
-
-            try {
-                const { object: extractedParameters } = await generateObject({
-                    model: aiModel,
-                    schema: z.object(schemaObject),
-                    prompt: mcpUtils.buildFinalExtractionPrompt({
-                        parameterExtractionPrompt,
-                        propertySchemaValues,
-                    }),
-                })
-
-                return {
-                    ...accumulatedParameters,
-                    ...extractedParameters,
-                    'auth': connectionReference,
-                }
-            }
-            catch (error) {
-                logger.error({ error }, 'FailedToExtractParametersFromAI')
-                throw error
-            }
-        }, 
-        Promise.resolve({ 'auth': connectionReference }),
-    )
-
-    const nonNullExtractedParameters = Object.fromEntries(
-        Object.entries(extractedParameters).filter(([_, value]) => !isNil(value)),
-    )
-    return nonNullExtractedParameters
-}
-
 function isOkSuccess(status: number) {
     return Math.floor(status / 100) === 2
 }
@@ -437,23 +333,7 @@ function trackToolCall({ mcpId, toolName, projectId, logger }: TrackToolCallPara
     }), logger)
 }
 
-type ExtractActionParametersParams = {
-    actionMetadata: ActionBase
-    toolPieceMetadata: McpPieceToolData
-    userInstructions: string
-    piecePackage: PiecePackage
-    platformId: string
-    projectId: string
-    logger: FastifyBaseLogger
-    mcpId: string
-}
 
-
-type InitializeOpenAIModelParams = {
-    platformId: string
-    projectId: string
-    mcpId: string
-}
 
 type TrackToolCallParams = {
     mcpId: string
